merge

+++ backend/order-service/src/main/java/com/example/orderservice/services/OrderService.java
@@ -2,10 +2,11 @@ package com.example.orderservice.services;
 
 import com.example.orderservice.controllers.clients.ICartClient;
 import com.example.orderservice.controllers.clients.IProductClient;
-import com.example.orderservice.core.CartOrderEntityMapper;
-import com.example.orderservice.core.OrderDTO;
-import com.example.orderservice.core.OrderMapper;
-import com.example.orderservice.core.ReqProductDTO;
+import com.example.orderservice.core.mappers.CartOrderEntityMapper;
+import com.example.orderservice.core.dto.OrderDTO;
+import com.example.orderservice.core.mappers.OrderMapper;
+import com.example.orderservice.core.dto.ReqProductDTO;
+import com.example.orderservice.core.exceptions.InsufficientQuantityException;
 import com.example.orderservice.core.exceptions.ResourceNotFoundException;
 import com.example.orderservice.entity.OrderEntity;
 import com.example.orderservice.entity.cart.CartEntity;
@@ -39,16 +40,21 @@ public class OrderService implements IOrderService {
         if (cartEntity.getProducts().isEmpty()) {
             throw new ResourceNotFoundException("Cart for user with id: " + user + " is empty");
         }
-        OrderEntity orderEntity = CartOrderEntityMapper.cartToOrderEntity(cartEntity);
+        long count = cartEntity.getProducts().stream().filter(p -> p.getAmount() > p.getProduct().getAmount()).count();
+        if (count == 0) {
+            OrderEntity orderEntity = CartOrderEntityMapper.cartToOrderEntity(cartEntity);
 
-        orderEntity.getProducts().forEach(p -> productClient.delete(new ReqProductDTO(p.getProduct().getId(), p.getAmount())));
-        cartClient.cleanCart(user);
-        repository.save(orderEntity);
+            orderEntity.getProducts().forEach(p -> productClient.delete(new ReqProductDTO(p.getProduct().getId(), p.getAmount())));
+            cartClient.cleanCart(user);
+            repository.save(orderEntity);
+        } else {
+            throw new InsufficientQuantityException("there is not enough product in stock for " + count + " items");
+        }
     }
 
     @Override
     public Page<OrderDTO> getPageForUserWhereOrderIsReady(String userEmail, Pageable page) {
-        Page<OrderEntity> orders = repository.findAllByUserEmailOrderByReady(userEmail,page);
+        Page<OrderEntity> orders = repository.findAllByUserEmailOrderByReady(userEmail, page);
         return orders.map(OrderMapper::convertOrderEntityToDTO);
     }
